Restore -no-undefined to *la_LDFLAGS.
This flag was removed as a temporary workaround for MacOSX for python, but
we now have a better solution.  The PYTHON_DEVEL macro supplies
PYTHON_LDFLAGS which gives us the required library so that all symbols
should be resolved on all platforms when linking UNDERSCORE_plplotcmodule
(a.k.a _plplotcmodule) and plplot_widgetmodule.

All symbols were resolved already for the java case without requiring any
extra libraries so it was never necessary to remove -no-undefined in the
first place for that case.

svn path=/trunk/; revision=5389