Implemented part of the equivalent of x20c demo in PerlDL.  The majority of the file is now just a boilerplate, but at least one call of plimage is working.  This puts the Perl binding pretty close to the C and C++ ones.
Although the plimage is poorly documented and only "half-official" in the
API, porting x20c proved to be an interesting exercise, exploring all the
power of PerlDL.  For instance, the cumbersome C routine for reading the
lena.pgm file in x20c.c was replaced by seven lines of simple code based on
the GD Perl module.  The file lena.png, which is read by x20.pl is also
added to the repository.

svn path=/trunk/; revision=5551