Dropped mention of editing plplot-config in the deb and rpm section.
Here is the reason: That idea won't work; compile will look in the wrong
place for headers, and the link will look in the wrong place for libraries
on the build machine.  Instead, if plplot-config is left as-is, it will do
everything right except put in the appropriate --rpath option for the build
machine (which is unlikely to be /usr/lib which is required on the users'
machines for debs and rpms).  That was the run-time problem I was concerned
about originally, but in fact it doesn't matter since /usr/lib will be
checked in any case for the PLplot libraries on the users' Linux machines,
and all the remaining required links are hard-coded into the PLplot library
with the correct /usr prefix values during the deb or rpm build.

svn path=/trunk/; revision=5558