Sort out two minor bugs.
* plP_gzbackw replaces plP_gzback so that plbox3 can store line width
information for subsequent plotting of the back planes for plsurf and plot3d.
Without this change, changes in the width of the contours were also occurring
in the back plane line width.

* Move the surface plcol1 call inside the z clipping logic (which means
colours won't be out of range for z values outside the clipped range).

Also changed logic so that surface contours were no longer necessarily the
background colour.  Instead, they are set by the user using plcol0. Example
8 from the various languages actually set this contour colour to yellow, and
I prefer it to the black colour so I will leave it that way.

svn path=/trunk/; revision=5578