Changes to the freetype routines to add text buffering support.
While adding freetype support to the wingcc driver I discovered that
"plRemakePlot()" caches line drawing commands, and things line
changes to the pen, but it does not cache text commands (well, it
does... sort of... as line drawing primitives). That meant that
every time a "plRemakePlot(pls);" was issued, the text would
disappear. To QUICKLY compensate for this limitation without major
changes to the API, these changes add a text cache to freetype.
These changes are really for interactive terminals that may get a
"redraw" command, and use plRemakePlot to refresh the display.

There is one new "public" function, one function-like macro, a
private function, and some changes to the FT structure to
accommodate the new cache.

The new cache is "activated" if "pls->plbuf_write" is set to one
before the freetype engine is initialised. Then, there are only
three more things the user has to do:
1) to the "plD_bop_[...]" function, add a call to
"pl_FreeTypeBOP();", which resets the text cache for the new page;
2) Add a "pl_RemakeFreeType_text_from_buffer(pls);" command after
your "plRemakePlot(pls)" command; and,
3) If there is a "resize" command, update a couple of freetype
internals so they know of the new size, i.e.
          if (FT)
            {
              FT->scale=dev->scale;
              FT->ymax=dev->height;
            }

To have a look at it "working" examine wingcc.c
----------------------------------------------------------------------

svn path=/trunk/; revision=5614