Changes to wingcc.c to get freetype support working.
This has involved adding a new backend into freetype and a local
text cache.

There are some minor caveats with the driver at present:
1) Occasionally the pen colours are getting messed up on redraws
when freetype text is being used. I am not sure why just yet;
2) if freetype is used without smoothing, sometimes the incorrect
pen colour is used for the text (the correct one is always used with
smoothing); and,
3) If a window is RESIZED, the label text wont change sizes
initially - all other text will though. BUT if you cause another
redraw of the window, for example, by violently moving it, putting
another window momentarily on top of it, or minimising it then
immediately maximising it again, then the text will be redrawn in
the correct size. Not sure if this is an API, freetype, or wingcc
problem.

----------------------------------------------------------------------

svn path=/trunk/; revision=5615