pltcl calls functions defined in libplplot.  It is linked to the tcl wrapper library libplplottcltkd which in turn is linked to libplplot, and libtool figures out how to resolve all the symbols in pltcl from this.  Also, at run-time for installed examples all is well since libtools sets the rpath options appropriately.
However at run-time for make check in the build tree there is what I believe
is a bug in libtool since it appears that although all the relevant
information is available in the relevant *.la files, the run-time loader is
not properly informed of the location of libplplot in the build tree.  You
must set LD_LIBRARY_PATH to point to the build-tree location of libplplot or
else install the libplplot library (which opens up the can of worms that if
you rely on the installed version, it might be out of date relative to build
tree changes you are testing).

To work around this bug in libtool for the make check case, I now explictly
link pltcl to libplplot. This change allows make check to work at run time
for all the tcl examples, but this redundant linking should not be necessary
and should be removed when the libtool "make check" bug is sorted out.

svn path=/trunk/; revision=5661