Compile java files in the build tree, create jar files, then install jar files. Just like we do for other languages.
There are two separate jar files so the user can play with and rebuild the
examples if they like. There is a Makefile in share/plplot/examples/java
for this - like the Makefile.examples for other languages.

Update test_java.sh script to use the new jar files. As long as everything
is in the installation directories set at configure time there is no need
to set the CLASSPATH.

Tested with gcj and with Sun JDK. Please check this on your system,
particularly if you are not using Linux.

svn path=/trunk/; revision=5699