Add a new Plplot java class. This effectively wraps the plplotjavac class in an object orientated manner, rather like the C++ plstream class. Each instance of a class is associated with a plplot stream. Loading of the java library and switching streams is handled automatically. This is what was always intended for the java wrapper I believe, but never happened with swig.
Update examples to use the new Plplot class. This removes most of the
warning messages generated by jikes. A few remain because the new class
implements the plplotjavacConstants interface to get the constant PL_*.

Since this is built on top of the existing bindings it is completely
backwards compatible. To add new functions you need to add them to
Plplot.java as well as to plplotcapi.i (in the swig directory).

svn path=/trunk/; revision=5719