Changed zoom & save setup to read option values from the resources database. This will allow users to set these according to taste from an app-defaults file, instead of always having to access the tk menu.  Also makes it nicer for applications built on top of PLplot to override.
Wrote from scratch an "unlist" proc and converted to use it everywhere.

Split off the code that determines the multiplier on pan operations, so that
it can be called from clients.  See get_scroll_multiplier for info.

Changed label_refresh and label_pop to allow the client to pop off the top
label (leaving an empty label field) if so desired.  After calling label_pop,
must call label_refresh before the effect is seen.

svn path=/trunk/; revision=5733