Put files in correct places inside plplot.jar, i.e. bindings/java/*.java and README.javaAPI go into plplot/core, and examples/java/x??.java and README.javademos go into plplot/examples.
This completely eliminates the need for the examples_execjavadir and
ourexecjavadir variables. This is definitely a Good Thing (TM), because
Automake *dir variables should only be used when files are actually
installed.  In the past, the files were installed in
$(datadir)/java/plplot/{core,example}/, the jar command was run in a
install-hook rule, and finally the files were rm -rf.

Oh boy, that was hackish, not to say weird! :-)

svn path=/trunk/; revision=5741