Changes to get make check working with the --srcdir option.
Seems a little inconsistent to have some of the test_xxx.sh scripts
generated from the corresponding .in files but not others. Also this means
that some scripts are in the source tree while others end up in the build
tree. Move the .sh file to .sh.in and alter cf/test.ac so they are all
generated by automake.

Also add the PLPLOT_LIB variable in test/Makefile.am so the examples can
find the font files in the source directory.

svn path=/trunk/; revision=5749