Change from HAS_pdl option and corresponding has_pdl conditional to enable_pdl option which can be set by user and corresponding enable_pdl conditional.
By default enable_pdl is turned off because pdl-2.4.2 (which has the
required interface to PLplot) has not been released yet.  But experimentally
you can try it if you have the pdl-2.4.1 version from Debian testing (which
also contains the PLplot interface), and good results are obtained in that
special case.

svn path=/trunk/; revision=5772