Two important changes to the configuration bootstrap script:
(1) replace the part of the script that runs the required autotools commands
in the correct order and with the correct options to bootstrap the PLplot
configuration system.  The replacement is just simply running the autoreconf
script as strongly recommended by the autotools documentation.

(2) remove the aclocal option processing.  Rafael kindly supplied this in
the old days in response to my need to run downloaded versions of the
autotools rather than system versions.  However, I don't need this any more
(Debian testing has virtually the latest autotools), and it may turn out
if/when we need this facility in the future that autoreconf handles this
properly, i.e., uses autotools files consistent with the downloaded
autotools version you have installed and specified by changing your PATH
appropriately.

N.B. I have discussed both these experimental changes with Rafael and he
said essentially try it and see.  So that is what I am doing.  They work
fine for me.  But we need more extensive testing.  If anybody has trouble
with them, I am mentally prepared to revert these changes.

svn path=/trunk/; revision=5817