Revert to explicit use of various autotools commands rather than the recommended autoreconf command.  It turns out for the particular way we use libltdl we need the libtoolize --ltdl option.  We might change our libltdl approach in the future which would allow us to go back to using autoreconf.
This is only a partial reversion of the last commit.  Compared to two
versions ago we still have the following changes:

(1) aclocal options can no longer be specified (I don't think we have need
of this, see previous commit message).

(2) we use relative path for cf directory which means macros in the cf
directory are accessed with m4_include rather than copied.  This makes the
tarball slightly smaller.  However, the real reason I am moving to this is
relative paths are illustrated in the documentation so it appears this is
the more usual way of doing things.

svn path=/trunk/; revision=5820