New function text2fci to parse font-changing commands in the middle of text.
New infrastructure for #<nnnn> and #<string/> escape methods of changing
font attribute values in the middle of strings.

For numerical escapes (#(nnn) and #[nnn]) and for greek escapes (#g), surround
results with an FCI to ask for symbol font, and an FCI to restore previous
font after the single unicode character.

Several other minor tweaks to commentary, more care in checking arguments,
etc.

svn path=/trunk/; revision=5896