Update plhrsh so that it uses the unicode_array method to communicate with ps.c.  (This change will require further changes with plfreetype.c.)
N.B. PLFLT xform[] = {1.0, 0.0, 0.0, 1.0}; is a guess that works for
examples 6 and 7, but there are interactions with whether plhrsh2(ch, x, y)
is called first or not, e.g., the characters were twisted by 90 degs if
plhrsh2 is NOT called first (i.e., for the TEST_FOR_MISSING_GLYPHS option).
So further work is needed to set up xform properly in the general case.

svn path=/trunk/; revision=5908