Added fast (but libgnomecanvas buggy) rendering.
When using polylines on the Gnome Canvas, you sometimes get the following
error message: 'attempt to put segment in horiz list twice'.  This is an
error in libgnomecanvas, because the polylines are valid.

The workaround in the gcw driver is to plot out the polyline as a series
of line segments.  The rendering in this case is slower.  This is not
usually a problem, but can become so for real-time displays (e.g., the
plplotcanvas_animation).  So, a new function,
plplot_canvas_use_fast_rendering(), has been added.

The default is to use the slower rendering, because this always works.
Fast rendering works most of the time, but sometimes produces error
messages and obviously wonky plots for 3D contour plots and vector plots.

A better solution will need to be found in collaboration with the gnome
team.

svn path=/trunk/; revision=6036