Added the check-sync.pl script for checking whether the PDL scripts are in sync with the C examples.  The script looks for a line containing the following pattern:
    SYNC: x??c.c n.nn

where x??c.c is the corresponding C file and n.nn is the CVS (RCS) version
of which the Perl script is currently in sync.  For instance, here is what I
have right now:

    $ grep SYNC x01.pl x02.pl
    x01.pl:# SYNC: x01c.c 1.39
    x02.pl:# SYNC: x02c.c 1.14
    $ ./check-sync.pl x01.pl x02.pl
    x01.pl: sync with x01c.c 1.39
    x02.pl: sync with 1.14, but x02c.c version is 1.16

Notice that, for performance sake, this script only checks the CVS version
in the working examples/c directory and not in the SF repository.  It is
then necessary to run cvs up before running this script.

This script could easily be generalized to all the other bindings by moving
it to examples/.  Would it be useful?

svn path=/trunk/; revision=6110