Adjust configure.ac and automake configuration files for the following:
(1) rename gcw-driver.c back to gcw.c  (required for static device build)
(2) Build of libplplotgnome2 in bindings/gnome2
(3) Reorder subdirectory build traversal so that libplplotgnome2 always
gets built before when it is needed in bindings/python/gcw, but after
libplplot is built.  In the dynamic device case, it also is built before
the devices.  N.B. the present ordering works for both static and
dynamic devices.

svn path=/trunk/; revision=6283