The default hrshsym should be zero (as it is for gd) since TrueType fonts with a rich assortment of symbols are normally available.
This means that by default no old-fashioned Hershey fonts for symbols will
be used unless there is no equivalent unicode symbol.

The only reason to have hrshsym set to 1 by default is the ps.c case where
the accessible symbols are quite limited because of the limitations of the
Type 1 fonts associated with that device driver.

svn path=/trunk/; revision=6494