Added autotools infrastructure for the f95 interface and f95 examples files that have been added and updated before.
Note, this largely completes my contribution to the f95 interface and
examples. I can now build and install the f95 interface and f95 example, and
build and execute (with a good-looking result) the installed f95 example on
an Ubuntu Breezy system with gfortran.  However, other platforms need
additional testing, and additional f95 examples have to be programmed to add
to the current lonely single example.  Also, there are two issues with the
generated *.mod files that still need to be discussed on list.

svn path=/trunk/; revision=6511