Results of a kind review of our autotools configuration by Ralf Wildenhues, an active autotools developer.
configure.ac: reorder so that programme checks in c++.ac, f77.ac, and f95.ac
are done before AC_PROG_LIBTOOL is executed in libtool.ac.  This order is
recommended in the autotool documentation and turns out to be essential for
the CVS version of libtool (the precursor to libtool-2.x) to work properly.
Also, use AM_PROG_CC_C_O (deal with compilers that don't accept the -c and
-o options simultaneously).  c++.ac: Call AC_PROG_CXX before testing the C++
compiler. f77.ac, f95.ac: Use AS_IF in strategic places. libtool.ac:
AM_PROG_LIBTOOL ==> AC_PROG_LIBTOOL .

Comparison between a build using the set of files before these and the
current set of files indicates on a Debian stable system, the only
difference was "-lm' is now used almost everywhere for linking.  We have had
trouble before with -lm not appearing when needed for some systems so this
change may be beneficial.  The installed examples for a heavily loaded
PLplot system with all language interfaces deployed built with no obvious
problems and gave identical results to the old system.  However, more
testing of this new autotools ordering is required on additional platforms.

svn path=/trunk/; revision=6568