#define RETSIGTYPE void.  (This is only a first approximation, and should be replaced by the appropriate test result for whether signal is int or void, but that test hasn't been programmed yet.)
Change potential #define value from nothing to 1 when the in-place
documentation (originally generated by autoheader from the documentation of
our autotools configuration files which might be incorrect) says to do so.
This change allows both the #if and #ifdef macros to be used, while before
we were restricted just to #ifdef for these particular #define's.

svn path=/trunk/; revision=6830