Improve the RPATH treatment.
(*) CMake build-tree RPATH's will be fine automatically if library
dependencies are found with find_library, but if PKGCONFIG is used, you have
to process the result afterward with find_library to put it in the form (not
the -L and -l form!) where CMake does proper interpretation of build-tree
RPATHS.

(*) The install-tree RPATH's have to be stated explicitly for every library.
There is no inheritance from dependent libraries.

svn path=/trunk/; revision=6869