Restore ARGS subcommand which is necessary for the deprecated exec_program command.  At some point we should follow the general suggestion in the CMake documentation to replace exec_program by exec_process but we will continue to use exec_program for now.
Fix a number of dependency issues.  The worst missing dependency that I
fixed was that plplot-symbols.txt depends on libplplot(d).  I implemented
that via a target dependency of a custom target which file depended on the
custom command which creates plplot-symbols.txt.  Fortunately, the target
dependencies seem to be processed before the file dependencies so the
fix seems to work.

Attempt to sort out the multiple invocation of pdfjadetex.  The previous
solution (adapted from the CMake FAQ) meant pdfjadetex was constantly
being reinvoked for each subsequent build attempt after the first.  The
current solution solves that issue, but I have to do more testing of various
scenarios to be sure it works properly in all cases.

svn path=/trunk/; revision=7133