Add plcalc_world, plgfci, plgvpd, plgvpw and plsfci (all part of the common API) to the octave bindings. None of these are (yet) tested in the standard examples.
Note: definition of PLUNICODE (used in plgfci / plsfci) is different to
that in plplot.h. The script which generates the octave wrappers does
not recognise uint32_t so this may cause problems on platforms where
unsigned int != uint32_t. In fact the same problem also exists with
PLINT, but in practice this doesn't seem to cause any reported
problems.

svn path=/trunk/; revision=7224