Fix order of default coordinates consistent with C version of example 20. Cropped image now displays correctly with non-interactive drivers.
PDL handles reading pgm files slightly differently to our home-grown
routines in x20c.c for example. It returns the maximum number of the
used colour, rather than the number of colours encoded in the header
of the image. Hard code this to 255 for the perl example. Results
now consistent with the C version.

svn path=/trunk/; revision=7278