Initial commit of FindPkgConfig.cmake.  This module contains the pkg_check_modules macro which is much more reliable and useful than the venerable pkgconfig macro which is now deprecated in CMake cvs.
Note, the PLplot CBS still uses pkgconfig, and it will take some effort to
switch over to use pkg_check_modules instead so this commit of
FindPkgConfig.cmake is just a preliminary enabling step.

Also note, this commit is only a temporary measure to allow PLplot to use
FindPkgConfig.cmake (currently in CMake cvs) until it is released as part of
a normal CMake release.

svn path=/trunk/; revision=7358