AWI for Bill Paxton.  Apply scale error fix suggested by Bill Paxton in 2004-09-30.  (Better late than never!  ;-) ).  I did not follow his comments (below) in detail, but I confirmed that the fix solved an obvious alignment problem now we have this device driver working again so that is why I am finally committing it.
Bill Paxton's comments on the scale:

"The pstex driver tells TeX what the unitlength is for the picture, and this
better agree with what the ps driver tells PostScript, or we'll get
misaligned text.  The ps driver sets up the unit length by doing a scale
with values from ps.h, XSIZE and XPSSIZE,  which gives 540/720 for unit
length in PostScript.  The pstex driver sets the unit length to be
72./25.4/pls->xpmm.  Where pls->xpmm has been previously set by the ps
driver to be YPSSIZE/LPAGE_X.  YPSSIZE is ENLARGE*YSIZE.  LPAGE_X is
(PIXELS_X/VDPMM).  VDPMM is DPMM*32.  DPMM is 4.  PIXELS_X is 32768.  YSIZE
is 720.  ENLARGE is 5.  Crank all this and you finally get a unitsize from
pstex of 0.201575 vs. unitsize frm ps of 0.200000.

When pstex is changed to calculate unitsize in the same way that ps does,
the registration problem goes away.  The change I made in pstex.c was in the
plD_init_pstex procedure:

fprintf(fp,"\\setlength{\\unitlength}{%fbp}%%\n", 72./25.4/pls->xpmm);

Becomes

fprintf(fp,"\\setlength{\\unitlength}{%fbp}%%\n", 1.0/ENLARGE);

Heaven only knows what this change will cause to break elsewhere, but
it seems to work for my test cases."

Works for me (AWI) as well.

svn path=/trunk/; revision=7441