The thin interface needs a C pragma for plfill (otherwise the thin interface plfill does not work correctly as shown by x12a).
The thick interface needs an Ada pragma for plfill in order to 'access it.

My brute-force solution is to make an Ada pragma call plfillada and an C
pragma called plfill.  This builds without problems and seems to work at
least for the thin interface (e.g., x12a), but Jerry B. should review this
solution.

svn path=/trunk/; revision=7442