The SF_cvs2svn_plplot.sh bash shell script contains the steps used to create svn dump files of the plplot and www CVS modules, to create a local svn repository from those dumpfiles, and to test that svn repository in a large number of ways.  N.B. after the tests were completed, the dumpfiles were uploaded to SF, and used by svn software there to create our SF repository. When that repository was dowloaded and compared with the local repository they were identical except for some non-essential items like dates.  Thus, all the tests that were passed by the local repository would necessarily also be passed by the SF version so no further testing of the SF repository was done except to use it.
Part of those local tests involved comparing ChangeLogs generated from CVS
and SVN repositories, and the xsl script cvs2cl.xsl helped format the CVS
ChangeLog in the identical way to the SVN ChangeLog created by the svn2cl
command.

Both SF_cvs2svn_plplot.sh and cvs2cl.xsl are specific to PLPlot and unlikely
to be used again, but I thought it was appropriate to commit them to the
"old" directory for historical interest about exactly how our CVS repository
was converted to an SVN repository and checked.


svn path=/trunk/; revision=7653