Replace ${CMAKE_CURRENT_BINARY_DIR}/massage$(EXEEXT) everywhere it occurs by ${massage_LOCATION} (where massage_LOCATION is determined by get_target_property on LOCATION).
The $(EXEEXT) was not correct in any case and should have been ${EXEEXT},
but even then there would have been known cross-platform problems with that
method, and (possibly?) it might not work for a "multitasking" make, e.g.,
"make -j2".  The LOCATION way of specifying dependencies is the most general
one and has proven to work for all platforms and also (at least for one
platform) for make -j2.

svn path=/trunk/; revision=7687