AWI for Jerry Bauck
Modified all Ada bindings so that user programs (and thus the PLplot
examples) no longer need to express a dependence on the thin binding. Thus,
user programs and examples no longer need to "with" or "use" PLplot_Thin,
and the namespace collision described in the last patch no longer exists
(for the user).

Moved some constant declarations out of the thin binding into the other
bindings to improve visibility for the user and to remove the need to
reference PLplot_Thin. There is now some duplication between the
higher-level bindings PLplot and PLplot_Traditional and the lower-level
PLplot_Thin in these constants, but the user doesn't see the duplication
since there is no need to look at PLplot_Thin.

Added two new files to the bindings directory, plplot_auxiliary.adb and
plplot_auxiliary.ads. This was necessary in order to make the PLplot_Thin
bindings disappear to the user. These files contain a few basic declarations
used by all of the bindings (which were formerly exported by PLplot_Thin)
and some utility procedures used by several of the examples.


svn path=/trunk/; revision=7741