Make all source files for the Ada bindings configurable using the CMake variables Ada_Is_2007_With_and_Use_Numerics and Ada_Is_Not_2007_Vector_Matrix_Declarations.
cmake/modules/ada.cmake defines the variables depending on an option set by
the user.

bindings/ada/*.ad? are the configured files whose configured results end
up in the equivalent directory in the build tree.

bindings/ada/CMakeLists.txt does the actual configuration of the source files
for the Ada bindings.

examples/ada/CMakeLists.txt is adjusted for the new location of the
(configured) source files for the bindings.


svn path=/trunk/; revision=7928