get-drv-info.c changed to dynamically load the device just like the code in plcore.c.  In particular, plGetDrvDir is called to obtain the devices directory.  That information is used so that the dynamically loaded shared object for the device is accessed with a full path name just like for the code in plcore.c.
This change eliminates potential nameclashes when using get-drv-info (strace
showed that libltdl was searching many different directories in the OS tree
for ps.la and ps.so in one get-drv-info test case before this change) and
the get-drv-info dynamic load should now work identically to the one in
libplplot.  This change may also fix the get-drv-info issue on Cygwin where
libltdl could not find ps.so when that file was just specified without the
path.


svn path=/trunk/; revision=7981