Large round of changes to make the use of char * / const char * more consistent in the API. Externally this affects the C and C++ language bindings. Internally there are also changes to f77, f95 and swig generated bindings. Now the only uses of char * in plplot.h are for functions which genuinely return or modify a string.
This is not strictly a binary API change as the const modifier in API 
calls is just telling the compiler that the function won't modify the 
string. This will have no impact on the user code. As such I have only 
upped the minor version number in the library SOVERSION for plplot and 
plplotcxx. I have tested on Linux and programs compiled with the old 
version link and run fine with the new version. Please test on other 
platforms. 

This patch also includes some internal tidying up to make things 
consistent with the updated API.

All C and C++ examples have also been modified consistent with these 
changes. 


svn path=/trunk/; revision=8033