Fixed indexing error in x16a.adb and xthick16a.adb that caused bad distortion on the fifth plot.
Fixed a rounding problem in the Ada versions of pltr1 and pltr2 in plplot_thin.adb: Ada converts floats to integers by rounding while C does so by truncation. Subtracting 0.5 - 10^-16 before converting seems to work at lest for 64-bit floats but in general there is no foolproof fix for this problem. This bug caused thousands of subtle differences between the Ada and C versions of Postscript files for Example 16.

svn path=/trunk/; revision=8511