Changes to tcl example 20 to ensure it works with tcl 8.4.
Fix some errors with quotes in pllab.

Avoid use of u flag in binary scan. This is only available with tcl8.5 
upwards. Many distributions still have 8.4 as default.

Fix a few other bugs with the interactive part of the code. 

The example works with non-interactive drivers (e.g. psc) but not with 
interactive drivers such as xwin.


svn path=/trunk/; revision=8729