AWI for Hezekiah M. Carty.
Substantial (at least factor of 4) speedups to pltr1 and pltr2.

Renames some of the OCaml-specific API functions to have a "pl" prefix for
consistency and to help avoid namespace collisions when the Plplot module is
opened in OCaml.

Force pl_setcontlabelformat to have the correct name (no "c_" prefix) under
OCaml.


svn path=/trunk/; revision=8738