Attempted to fix the problem with the callback in Ada examples 19. When C calls an Ada subprogram while passing an unconstrained array, even if C calling conventions are in force, there is no guarantee that anything will work since C has no way of passing the array start to Ada and Ada will just take a guess (worst case) or issue a warning or error (but not GNAT) or naively look up a bogus number. I have thus used a constrained array of bounds 0 .. 2000 for the mapping function called by plmap and plmeridians, in the API called mapform.
There is a discussion of this issue at http://groups.google.com/group/comp.lang.ada/browse_thread/thread/0e5a3abec221df39?hl=en#.


As an unrelated set of edits (included here because SVN write was down for maintenance earlier):

Removed the requirement of with-ing PLplot_Auxiliary in user programs. PLplot_Auxiliary is still available as a convenience to the user but with reduced functionality, providing only vector and matrix maximum and minimum calculations. It is used in several of the examples as well as the thick bindings.

svn path=/trunk/; revision=8749