Implement familying.  I didn't track potential memory management issues in the code (which can be quite complicated), but I did test the result with valgrind for multipage examples 8 and 9 and got an absolutely clean valgrind result.
The current situation (just like before familying was enabled) is if the
-fam option is _not_ used this device produces bad svg results (repeat <svg>
elements corresponding to each page which is not allowed by the svg standard
which is only a single-page standard). These bad svg results completely
break the imagemagick viewer, cause errors and warnings for other viewers
(while they pick an arbitrary page to view), and invalidate the results at
the w3c validator.  Of course, none of this is an issue now if the user
remembers to always use the -fam option for multipaged plots.

ToDo: to avoid this issue when -fam is not used for multi-page plots do file
output only for the case of the first page similar to what is currently done
for the gd devices and the single-page subset (pngcairo and svgcairo) of the
cairo devices. For all these cases we should issue a reminder message to use
-fam whenever any pages are being skipped.


svn path=/trunk/; revision=8830