Scale coordinates so that internal PLplot virtual pixel coordinates are larger (PIXELS_X x PIXELS_Y or 32768x24576) than the normal external coordinates used for the svg file.  This change improves the PLplot internal precision sufficiently to solve the hidden line errors that occurred previously for example 18.
This change left some continued lack of smoothness in the results (although
no known hidden line problems).  This additional issue was solved by storing
the above scale factor and all the external coordinates written to the svg
file as doubles.

ToDo.  Use explicit formatting on the coordinates written to the file to
round values to say 0.01 pixels.  This small a rounding error should not
significantly compromise the smoothness and should substantially reduce the
size of the svg files that are generated.


svn path=/trunk/; revision=8834