Apply variable-opacity black shade to left side of plot with plshades.  This commit illustrates that plshades is not adequate for this task of approximating a gradient.  The various rectangles overlap each other slightly given twice the opacity at the edges and therefore an extremely uneven effect.  If you use lots of rectangles (say 1001 points in clevel), then many different edges overlap simultanously given much higher opacity that blackens the whole left-hand side of the plot much too much which makes this method unacceptable even if you were willing to accept the file bloat caused by it.
Standard example 30 also shows the deficiencies of plshades (which is the
best method PLplot has) for approximating gradients.

The purpose of this commit is to illustrate the PLplot gradient problem.
There is currently no PLplot solution for this issue. Therefore, this
attempt at approximating a gradient with PLplot will be removed in the next
commit.


svn path=/trunk/; revision=8923