Commit of a modified form of a patch submitted by Hezekiah M. Carty to make sure that the stream variables plsc->vpwxmi, plsc->vpwxma, plsc->vpwymi, and plsc->vpwyma retain the exact values specified by the user's call to plwind so those exact values are delivered back by plgvpw.  An internal function called plP_xgvpw is now used to deliver the (slightly) extended versions of these variables that PLplot uses internally so that the end limits will be on the graph.
The net result of this change is that C/C++ users who access the raw stream
variables and all language users who use the plgvpw function to access them
will get the exact values input by the user in plwind rather than the
slightly extended values.  However, after some discussion, Maurice decided
this downside of the change was okay because this change will affect a
relatively small percentage of our legacy users, and I am certainly willing
to go along with his conclusions on this.

The examples do not use either the raw stream variables or plgvpw so they
should come out identically to before.  This happens for *.psc, *.psttfc,
*.png, and *.svg.  

In contrast to these voluminous expected results *.pscairo has differences
for examples 7, 12, 23, 24, and 29 and *.pngcairo has differences for all
examples.  It turns out the *.pngcairo differences are due to timestamps
(see list discussion for the evidence) so that is okay, and the only worry
is the *.pscairo differences.  Those turn out to be something to do with
fonts and probably nothing to do with the present changes which is why I am
going ahead and making this commit.


svn path=/trunk/; revision=8942