Implement configured versions of style_liquid.css and style_static.css. So far only the sizes are configured, but the results for both configured styles work well when hand-copied to the website and also validate (with the colour warnings).
ToDo: 

(1) configure the colors (which should get rid of the above colour warnings).

(2) replace current Makefile logic (which is limited to just the source tree
for copying to the website) with CMake logic (which can handle the
build-tree location of the configured style_liquid.css and style_static.css
when copying results to the website).  Until this change is completed,
uploading the website automatically will not work and the configured results
must be uploaded by hand.


svn path=/trunk/; revision=8959