Fix a couple of compile warnings with gcc4.3.2.
For constant strings it is better to use fputs rather than fprintf. This 
prevents inadvertently or maliciously including format characters such 
as %d in the string which would cause fprintf to fail.

Also, on 64-bit platforms size_t may not be an int so explicitly cast 
the value to an int to prevent mismatch with %d format. Newer glibc 
versions include a z length modifier which signifies size_t, but unless 
this is widely available it is likely only to cause additional problems. 


svn path=/trunk/; revision=9110